requirements.txt
# Core dependencies
pyspark==3.3.0
xgboost==1.6.2
scikit-learn==1.1.2
pandas==1.5.0
numpy==1.23.3
boto3==1.24.71
mlflow==2.0.1
matplotlib==3.6.0
plotly==5.10.0
python-dotenv==0.21.0

# API and web dependencies
fastapi==0.85.0
uvicorn==0.18.3
gunicorn==20.1.0
pydantic==1.10.2

# Testing
pytest==7.1.3
pytest-cov==3.0.0

# Linting and formatting
black==22.8.0
flake8==5.0.4
mypy==0.971